TestItemModel, moved to order service
