+++ Anlab.Mvc/Models/Order/OrderEditModel.cs
@@ -3,12 +3,13 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Anlab.Core.Domain;
+using Anlab.Core.Models;
 
 namespace AnlabMvc.Models.Order
 {
     public class OrderEditModel
     {
-        public TestItem[] TestItems { get; set; }
+        public TestItemModel[] TestItems { get; set; }
         public Anlab.Core.Domain.Order Order { get; set; }
 
         public string DefaultAccount { get; set; }
